feat:CTTIIDPS-7632 add retry status code 529 and dates in resource log

+++ cloud_development/app/domain/AzureSqlMetric.py
@@ -13,7 +13,7 @@ class AzureSqlMetric:
     name: str
 
     tablesQuantity:int
-    tablesDenormalized:int
+    tablesDenormalized:Decimal
     tablesDenormalizedPoints:Decimal
 
     topConsumptionQueries:int
